fix: don't qualify paths of registered types
The previous commit added qualification to paths of registered types.
However, while types defined in Carp support qualification (definition
within a module) thanks to special logic, this logic is not yet
implemented for registered types, leading to errors if their paths are
qualified in the type environment. This commit fixes the behavior and
adds a TODO to support this for registered types.