fix: don't type check untyped forms in set!
This is a kludgey temporary fix for #1179. The problem is that set!
occasionally receives untyped forms. In particular, this seems to happen
when the value of the form is determined by a command such as `length`.
This caused hangs when evaluating `Dynamic.mod` in particular.

For now, we just treat these like dynamic values and don't worry about
typechecking them, which seems to do the trick.