fix: fix parenting chain when handling recursive lambdas
Recent commits to support let-rec incorrectly ordered the recursion and
function environments, resulting in captured variables not being
captured. This change alters the parenting relationship between an inner
let env and the recursion environment needed for recursive lambdas to
fix the issue. (in particular, see examples/nested_lambdas for the case
this impacts).