refactor: move form validation into a separate module
Historically, the evaluator both checked that incoming Carp forms were
well formed and evaluated them when this was in fact the case. This
complicates the evaluator by giving it multiple responsibilities.

This commit introduces a new module, Forms, that is purely responsible
for the validation of the well-formedness of Carp expressions. The
evaluator calls this module to validate forms before evaluation. This
enables us to modularize our code a bit better. In addition, this commit
introduces pattern synonyms for common forms, simplifying pattern
matching in some cases.

This commit also contains a fix for a previously unreported error in the
evaluator that introduced hangs on forms like `((def foo 5) 5)`.