feat: Add an initial frontend for protocols
This commit adds frontend support for two new primitives, `defprotocol`
and `instance`.

defprotocol defines a new protocol, a collection of zero or more
interfaces. Protocols implement a form of subtyping whereby one can
assert that a given type implements a number of interfaces. Types that
implement a protocol are called `instances` and are added using the
`instance` primitive.

This commit does not yet add any logic for enforcing protocols, only the
frontend code for adding them to a context. Initially, we will only
support protocols in the specification of interfaces. A protocol may be
used in place of any type or type variable.