feat: respect protocols in the type system
This commit finalizes an initial implementation of protocols by updating
the type system to treat them as instances of constrained polymorphism
(that is, like var types but more restrictive).

Protocols are a form of type constraint whereby only types that
implement a set of interfaces may be used in the place of a protocol.

Here's the gist of how they work:

To define a protocol, a user enters `(defprotocol <name>
[interfaces...])`. `defprotocol` adds a fresh protocol to the type
environment, inhabited by no types.

Users then register types with the protocol using `(instance <protocol>
<type>)`. The type will only be registered if it appears in at least one
position of the signatures of the protocol's interface's
implementations. That is, if there is no existing implementation of all
<interfaces> for the type, it cannot be added to the protocol.

Users can then reference protocols in type signatures by prefixing their
names with `!`. For example, `(sig foo (Fn [!Stringable] String))`.

When we type check protocols, we update the state of types based on the
current state of the type environment (since a protocol's members may
have changed). If the type being passed in the place of a protocol is a
member of the protocol, it successfully type checks. Otherwise, it fails
to type check.

Protocols can also be used for arbitrary "tagging" or "subtyping" over
types by defining empty protocols. If a user wants to simply specify a
type union, they can call:

`(defprotocol Empty)` which will define a protocol that has no
interfaces. Any type may be added to this protocol.